size_t is not equivalent to int on 64-bit platforms.
Pointy hat to:	edwin@

+++ usr.bin/calendar/parsedata.c
@@ -39,9 +39,9 @@ __FBSDID("$FreeBSD$");
 static char *showflags(int flags);
 static int isonlydigits(char *s, int nostar);
 static const char *getmonthname(int i);
-static int checkmonth(char *s, int *len, int *offset, const char **month);
+static int checkmonth(char *s, size_t *len, size_t *offset, const char **month);
 static const char *getdayofweekname(int i);
-static int checkdayofweek(char *s, int *len, int *offset, const char **dow);
+static int checkdayofweek(char *s, size_t *len, size_t *offset, const char **dow);
 static int indextooffset(char *s);
 static int parseoffset(char *s);
 static char *floattoday(int year, double f);
@@ -771,7 +771,7 @@ getmonthname(int i)
 }
 
 static int
-checkmonth(char *s, int *len, int *offset, const char **month)
+checkmonth(char *s, size_t *len, size_t *offset, const char **month)
 {
 	struct fixs *n;
 	int i;
@@ -822,7 +822,7 @@ getdayofweekname(int i)
 }
 
 static int
-checkdayofweek(char *s, int *len, int *offset, const char **dow)
+checkdayofweek(char *s, size_t *len, size_t *offset, const char **dow)
 {
 	struct fixs *n;
 	int i;
